Added usings of argparser.

+++ src/filesorter/sorter.py
@@ -1,65 +1,101 @@
 """
 """
-
+import functools
 import json
 import os
+import queue
 import sys
 import time
 
 from pathlib import Path
 
-
 import logger.config as config
 import logger.logger as Logging
 
-from logger.logger import logger,formatter_result
+from logger.logger import logger, formatter, formatter_result
 
-config.SCRIPT = __file__
 # print(__name__)
-if      __name__ == "__main__"\
-    or  __name__ == "filesorter.sorter"\
+if      __name__ == "__main__"          \
+    or  __name__ == "filesorter.sorter" \
     :
-    import filesorter._argparser as _argparser
+    import filesorter._argparser as argparser
     # Override default config
-    config.LOG_FILE_TRUNCATE = True
-    Logging.init()
-    Logging.truncate()
+    if not config.SCRIPT:
+        config.SCRIPT = __file__
+        config.LOG_FILE_TRUNCATE = True
+        config.PROFILING = True
+        Logging.truncate()
+        Logging.init()
 
 if config.PROFILING:
     import cProfile
 
-from filesorter._executors import registry
+from filesorter._executors import registry, TASK_SENTINEL, RESULT_SENTINEL
 from filesorter.filters import Filter, Filters
 from filesorter.task import Task
 
 
 
-def sort_targets(path_to_target, executor, settings):
+def timer(func):
+    """Print the runtime of the decorated function"""
+    @functools.wraps(func)
+    def wrapper_timer(*args, **kwargs):
+        start_time = time.perf_counter()    # 1
+        value = func(*args, **kwargs)
+        end_time = time.perf_counter()      # 2
+        run_time = end_time - start_time    # 3
+        print(f"Finished {func.__name__!r} in {run_time:.4f} secs")
+        return value
+    return wrapper_timer
+
+
+def sort_targets(args, executor, settings):
     # Make path list
-    if isinstance(path_to_target, str):
-        pathes = path_to_target.split()
-    elif isinstance(path_to_target, list):
-        pathes = path_to_target
-    elif isinstance(path_to_target, Path):
-        pathes = list(path_to_target.name)
+    if isinstance(args.directories, str):
+        pathes = args.directories.split()
+    elif isinstance(args.directories, list):
+        pathes = args.directories
+    elif isinstance(args.directories, Path):
+        pathes = list(args.directories.name)
     else:
-        raise ValueError(f"{path_to_target} value error.")
+        raise ValueError(f"{args.directories} value error.")
     
     # Remove duplicates
     pathes = list(dict.fromkeys(pathes))
 
+    tasks = []
     for path in pathes:
-        filters = Filters(path,_argparser.args.keep_empty_dir)
+        filters =   Filters(
+                        args.destination,#path,
+                        None,
+                        args.keep_empty_dir,
+                        args.normalize,
+                        args.overwrite,
+                        args.use_names,
+                    )
         for name, setting in settings.items():
             filters +=  Filter(
                             filters,
-                            name,
+                            name,#args.destination / name,
                             setting["extensions"],
                             setting["functions"],
                             setting["use"] if "use" in setting else ""
                         )
+        if args.name:
+            filters +=   Filter(
+                            filters,#path,
+                            args.name,
+                            args.extensions,
+                            args.functions,
+                            args.ext_executor,
+                        )
         task = Task(path, filters, executor)
+        tasks.append(task)
+
+    for task in tasks:
         executor.submit(task)
+    # else:
+    #     executor.submit(TASK_SENTINEL)
     return
 
 def load_settings(path = None):
@@ -71,15 +107,13 @@ def load_settings(path = None):
         dir2ext = json.load(settings)
     return dir2ext
 
-def main():
-    
+def sort(args):
     profile = None
     if config.PROFILING:
-        profile = cProfile.Profile()
-        profile.enable()
-        start = time.time()
+        # profile = cProfile.Profile()
+        # profile.enable()
+        start = time.perf_counter()
 
-    args = _argparser.parse_args()
     if args.settings:
         settings = load_settings(args.settings)
     else:
@@ -116,9 +150,11 @@ def main():
             }
         }
 
-    executor = registry[_argparser.args.executor]()
+    executor = registry[args.executor]()
+    executor.start()
 
-    sort_targets(args.directories, executor, settings)
+    sort_targets(args, executor, settings)
+    executor.submit(TASK_SENTINEL)
     executor.join()
     executor.shutdown(True)
 
@@ -126,18 +162,32 @@ def main():
         handler.setFormatter(formatter_result)
     # time.sleep(3)
     logger.info("<Result")
-    try:
-        while result := executor.results.get_nowait():
+    results = executor.get_results(False)
+    for result in executor.lresults:
             logger.info("\t" + str(result))
-    except Exception as e:
-        pass
+    # try:
+    #     while result := executor.get_results(False):
+    #         logger.info("\t" + str(result))
+    # except Exception as e:
+    #     pass
     logger.info(">")
 
-    print()
+    for handler in logger.handlers:
+        handler.setFormatter(formatter)
+    # print()
 
-    if config.PROFILING and profile is not None:
-        profile.disable()
-        profile.print_stats(sort="cumtime")
+    if config.PROFILING:
+        if  profile is not None:
+            profile.disable()
+            profile.print_stats(sort="cumtime")
+        end = time.perf_counter()
+        delta = end - start
+        logger.info(f"Total execution time: {delta}")
+
+def main():
+    parser = argparser.get_parser()
+    args = parser.parse_args()
+    sort(args)
 
 
 if __name__ == "__main__":
